complete authentication with jwt
