+++ server/src/controllers/likeControllers.js
@@ -1,0 +1,38 @@
+import { Post } from "../models/Post"
+
+
+// like a post
+const post_like = async (req, res)=>{
+	const id = req.body.postId
+
+	Post.findByIdAndUpdate(id,{
+		$push:{likes:{postId:id, userId:req.user.id}}},{
+			new:true
+		}).exec((err, result)=>{
+			if(err){
+				return res.status(422).json({err:err})
+			}
+			else{
+				return res.status(200).json({message:"post liked!", result})
+			}
+		})
+}
+
+//unlike a post
+const post_unlike = (req, res)=>{
+	const id = req.params.id
+	Post.findByIdAndUpdate(id,{
+		$pull:{likes:{postId:id, userId:req.user.id}}},{
+			new:true
+		}).exec((err, result)=>{
+			if(err){
+				return res.status(422).json({err:err})
+			}
+			else{
+				return res.status(200).json({message:"unliked post!", result})
+			}
+		})
+}
+
+
+export {post_like, post_unlike,}
