+++ server/src/models/Contact.js
@@ -1,0 +1,16 @@
+import {mongoose} from "mongoose"
+
+const contactSchema = new mongoose.Schema({
+	sender:{
+		type:String
+	},
+    name:{
+        type:String
+    },
+	message:{
+		type:String
+	}
+},{timestamps:true})
+
+
+export const Contact = mongoose.model("Messages", contactSchema)
